Create README.md
Anurag Verma is a talented front-end developer currently pursuing a Bachelor of Engineering in the IT branch. With a keen eye for detail and a strong foundation in HTML, CSS, and JavaScript, he excels in crafting visually stunning and user-friendly interfaces. Anurag's passion for web development drives him to explore emerging technologies and design trends, enabling him to create cutting-edge websites and applications. His dedication to continuous learning and academic pursuits makes him a promising asset in the ever-evolving world of front-end development.